Start on websocket renderer

+++ render/render.go
@@ -1,7 +1,7 @@
 package render
 
 type Renderer interface {
-	Init(w, h int) error
+	Init(w, h float64) error
 	StartFrame() error
 	FinishFrame() error
 	DrawAt(x, y float64, d Drawable) error
